+++ src/commands/serve-preact.js
@@ -42,7 +42,7 @@ function buildConfig(args) {
   }
   else {
     // Use a render shim module which supports quick prototyping
-    config.entry = [require.resolve('../preactRunEntry')]
+    config.entry = [require.resolve('../render-shims/preact')]
     config.plugins.define = {NWB_QUICK_MOUNT_ID: JSON.stringify(mountId)}
     // Allow the render shim module to import the provided entry module
     config.resolve.alias['nwb-quick-entry'] = entry