+++ src/commands/serve-react.js
@@ -36,7 +36,7 @@ function buildConfig(args) {
   }
   else {
     // Use a render shim module which supports quick prototyping
-    config.entry = [require.resolve('../reactRunEntry')]
+    config.entry = [require.resolve('../render-shims/react')]
     config.plugins.define = {NWB_QUICK_MOUNT_ID: JSON.stringify(mountId)}
     config.resolve = {
       alias: {